Rewrote AutoGCTRM and File Placement Code
AutoGCTRM and FIle Placement Changes:
Reconfigured the constants that drive file placement and AutoGCTRM, and updated the way those processes work.
This program now *always* produces its output in the "Code_Menu_Output" folder. It will then offer to copy its output over existing files (after backing up the originals) if they can be found in the build directory specified in CodeMenu.cpp.
If the program successfully backs up and replaces an existing Code Menu ASM file, and it is able to detect the files necessary to run GCTRM (the associated executable as well as the relevant .txt files for your codeset, eg. RSBE01.txt and BOOST.txt), it will offer to run that as well.

Other Changes:
- Separated out several utility functions into their own file to tidy things up a bit.
- Changed the default Build Directory such that the program will expect to be *within its own folder WITHIN the main Builds folder (eg. the "Project+" or "LegacyTE" folder)*.
- Spruced up log output to standardized formatting and be a bit more descriptive of the produced Menu's configuration.