Added Override Arguments to allow Handsfree Use
Added parsing for a set of 4 arguments. In order, these are:
- CMNUOverride
- ASMOverride
- GCTOverride
- PressButtonToCloseDisable

The first three can be set to either 1 or 0 to force the associated decision to process a value of "Yes" or "No" respectively, or they may be set to "-" to leave the interactive choice intact.

The final argument can be set to 1 to remove the need to press a key to close the program after it finishes running.

Also incremented version number to v1.0.2.