Setup For Options Output Implementation
The plan here is that, when building the menu, we'll also generate an XML which maps out all of the toggles, options, and other interactable bits in the menu. People will then be able to change what default values are selected in the menu by editing the file, and when they rebuild the menu, the program can use it to decide defaults on the fly.