Consolidated Options File Code
As a result of this reorganization, we no longer need the externed variables, and we can condense most of the functions related to the process to just one.

Additionally, since everything is localized to a single function call, we no longer need the node stack, which means we no longer need the stack library.