Rough Pass on Setting Base Character List From XML
Added a "characterListVersion" option to the menu options xml document, which you can change to decide which version of the base character list to use for the program. The code for this is kinda all over the place right now, I'm going to need to come back and reorganize all of this, but it appears to be in working order.