Re-implemented CSS Roster Switcher ASM
Previously, the code to drive the character switcher in-game was an external ASM file that you'd manually include in your GCTs, it's now been rewritten using the provided PPC Assembly bindings.
The code will only be included if the CSS Roster Version line is added to the code menu, which will itself only be added if more than one roster is defined when the menu gets built. Additionally, the generated code will automatically account for every filename specified in "ROSTER_FILENAME_LIST", no modification to anything necessary.