Simplified CSSChange ASM Code
Removed a couple unnecessary steps in the code. We don't need to use CSS_VER_LOC to find the location of the CSS Version line, we have that calculated and stored in CSS_VERSION_SETTING_INDEX already.

We also don't need to explicitly set the CSS Roster Path's address, then separately move past the folder portion; we can do that in one move since we know the values of both of those as well.