Brought In lavaOutputSplitter to Simplify Logging
Previously, all log messages had to be written twice: once to the log file, and a second time to cout. This was clunky and prone to mismatched text, so I've introduced another project I've been working on which makes it possible to output to both streams with a single statement.