Re-Implemented V1 Theme Funcs
Firstly, all of the V1 funcs have been changed to use a common body function to do the bulk of their work (editFilepathConstant()). All the individual functions are responsible for now is specifying the directory, base filename, path register, and path addresses for their respective menu files.

Additionally, that reusable body has been reworked a bit from the old process. The old approach relied on knowing where the filenames started in the filepath string constants, which was inconsistent between vanilla P+ and P+Ex. In this approach, we start by writing in the relevant paths *with* leading slashes to ensure they exist, then we proceed with the old approach of just writing in the necessary prefix as needed. Also, since we're writing in the strings using the default prefixes to start with, we don't need an explicit case in our IF block to handle that case!