Completely Reworked Theme Prefixes to be Per-File
Previously, any given theme was defined using a singular prefix, which applied to all files within that theme. This made things rather inflexible, in that any given theme *had* to affect every theme-able file, and thus that there needed to be a copy of each theme-able file with the corresponding prefix or else the game would crash. This was especially nasty in that, if a build had previously set up themes, and then updated to a newer version of this program which supported more theme-able files, that build would crash upon loading any newly affected files because they wouldn't exist.

With this rework, themes are defined within an XML file rather than a plain TXT file, in which each theme-able file may have its prefix defined individually. This allows themes to selectively affect some files and not others, paving the way for themes that for instance only affect the CSS, or only affect the main menu, for instance. Additionally, this solves the updating problem described earlier: any file which doesn't have a prefix defined will use the default prefix for that file! This also means that theme declarations may omit files they don't care to affect if they choose, as they'll simply be filled in as default by the program.

Additionally, temporarily omitted the V2 functions while I rework them to fit this new system.