Non-EX Builds Now Support Loading External Themes
Previously I'd set this up in the same way the Roster and Character loader settings were done, both of which only really make sense for EX Builds. That isn't the case with themes, however; those should work just fine on non-EX P+ builds (*any* Brawl build with a Code Menu, really).

Also changed the expected filename for the Theme input file accordingly, and deleted the old TXT file in the repo as well.