Reverted Theme File Name Change
The theme file being the only input file without the EX at the beginning is kinda ugly, and potentially un-intuitive given that all other input files (both in this, and in lKHM) follow the "EX_" convention. I'll just need to make it adequately clear in the readme that non-EX builds can take advantage of the theme features as well.