Began First Pass on New Code Naming Convention
Began implementing a new convention for naming codes: each code's name will start with "[CM: _______] ", where "CM" stands for "Code Menu", and the blank is the name of the .cpp file the code belongs to. My hope is that this will more clearly identify code menu codes in the "codeset.txt" files produced by GCTRM.

Additionally, I've added blurbs for a few codes that had relevant comments associated with them.