Initial Implementation for HUD Recoloring
Currently, this is set up to override the frame in a given menu object's CLR0 that gets applied, which allows us some control over what color those objects are. This approach mostly works, but because more objects go through the function I hooked than just those which are recolored based on player slot, I may end up needing to rewrite this to be less broad.

I've had a try with the damage percent colors at implementing a way to avoid overwriting the frame on functions I want to leave untouched, but needing to do it that way seems less-than-ideal.