Initial Shot at Trigger Flag Implementation
I think ultimately this isn't going to actually work, unfortunately. The plan initially was to just do a hook and set r11 to a specific value on every function call that I wanted the override to trigger on; but realistically I don't think that's gonna work. It just scales too poorly.