Implemented AND, ANDI(S), OR, and NOR Output
Additionally, implemented two of the main string conversion functions needed for the Op Code 31 instructions.

Also ensured that argumentStartBits vectors are the same length as those returned by splitHexIntoArguments, so that the secondaryOpCodeArgumentIndex field always indexes to the same argument in either of those vectors.

Lastly, instructions' "canonForm" fields are now automatically populated, and thus no longer need to be manually defined.