Implemented CSS and Results Name Transparency Code
This ensures that, when the Color Changer lines are included in the menu, the names for characters on the CSS and Results screen are made transparent, so that they can borrow their colors from the menu elements behind them.

Additionally, ensured that the code for changing the color of the Random CSP on the CSS is disabled properly when the color changer lines aren't enabled. Was already done for the other codes in here, that it wasn't on this code was a mistake.

Lastly, fixed inaccurate comment text on the checks for whether or not the Color Changer lines are enabled or not.