Implemented SLW, SRAW, and SRAWI
Additionally, changed my approach a bit with the opcode 31 instructions. There are certain instructions which use the highest bit of the secondary opcode starting at bit 21 as an argument for triggering an Overflow flag; which I was originally treating as its own argument in the instruction. Dedicating that bit to that argument though meant that any other instructions which had secondary opcodes which used that highest bit couldn't be correctly parsed, since when we split the operation's hex back into its constituent arguments we'd be cutting off the highest bit of its secondary opcode, and thus unable to identify it. Way around that is, rather than dedicating that bit to just that one purpose, we just explicitly define those Overflow versions of the relevant functions, and treat that 1 as part of their secondary opcodes.