Impl. System for Marking Sec Op Args in Layouts
This was implemented primarily to make it simpler to implement the "getTestHex" function, which generates a hex encoding for the given instruction in which every argument has the value 1, respecting any reserved argument values. This is useful for testing purposes primarily, it makes it a lot more apparent when a layout has its arg positions incorrectly defined (eg. this helped identify an issue where the CMPW layout had its RC arg set at bit 30 instead of 31).