Impl. Float Status and Control Reg Instructions
Includes MCRFS, MFFS, MTFSB0, MTFSB1, MTFSF, MTFSFI.

Additionally, implementing these has demonstrated a flaw in the process used to pull the secondary op-codes from instruction hex in the "convertInstructionHexToString" function. Essentially, because we were storing the possible start bits and lengths for secondary op-codes in a plain, unsorted vector, the order in which we used those regions to try to pull secondary op-codes from the incoming instruction hex was entirely dependent upon the order in which they were defined. This was a problem because, in the event that we defined a region, then afterwards defined a different region which entirely contained that previous region, we would end up checking that effective "sub-region" first, because it was defined first. This could lead to situations where reading that "sub-region" could cause us detect a secondary op-code match on an instruction whose secondary op-code *does* match that sub-region, but *does not* match the entire intended secondary op-code, thus giving us an incorrect match.

For example, in the case that revealed this to me, we have the "mcrfs" instruction, alongside the "fcmpu" instruction, both of which use primary op-code 63 (the double float arithmetic group). Importantly here, "mcrfs" uses secondary op-code 64 (0b1000000), while "fcmpu" uses secondary op-code 0 (0b0). That much is fine, but where the problem comes in is in the fact that there are 2 possible regions the secondary op-code can be in for this primary op-code group: either bits 26-30 (inclusive), or bits *21*-30 (inclusive). Because I had defined the 26-30 region first, when trying to determine which instruction a bit of instruction hex was referring to, we'd first pull bits 26-30, and see if that matched any of the catalogued instructions for this primary op-code group. In the case of an "mcrfs" instruction, that'd have us pull bits 26-30 from its hex, which would yield 0b00000 (or, simply, 0) as its proposed secondary op-code; a secondary op-code for which a match *does* exist: "fcmpu". As a result, the old setup incorrectly identified "mcrfs" instruction hex as "fcmpu" instructions, which is obvs a bad time.

The fix for this is rather simple actually: we just need to ensure that we always try specified secondary op-code regions in such a way that we never try a region before another region which wholly contains it (eg. we can't check bits 26-30 before we check bits 21-30, cuz 26-30 is contained in 21-30). One way we could achieve that simply would be to just change the order in which we define things such that we always define larger regions before smaller ones; but I didn't like the idea of solving the problem through convention rather than through implementation. *So*, as a solution, the container for proposed secondary op-code starts and stops has been changed from the old un-sorted vector, to a sorted set. Critically, this set is sorted first in ascending order by starting bit (a region which starts earlier than another region by definition cannot be wholly contained by that other region), and then in *descending* order by length (a larger region cannot be wholly contained by a smaller region). We can then check through the proposed regions in sorted order, guaranteeing that we never check a "sub-region" before other regions that contain it, (regardless of the order in which they were specified), eliminating the issue!

Lastly, I should note that this issue actually affected more than just "mcrfs". It additionally broke output for:

- "psq_lux" (38, 0b100110) and "psq_stux" (39, 0b100111), which would be incorrectly read as "psq_lx" (6, 0b000110) and "psq_stx" (7, 0b100110) due to another set of instructions in the paired single instruction group using bits 26-30 for the secondary op-code.
- "fcmpo" (32, 0b100000), which incorrectly read as "fcmpu" (0, 0b000000) due to another set of instructions in the double float instruction group using bits 26-30 for their secondary op-code (same as "mcrfs" lol).

This issue is completely fixed for those instructions as well thankfully; real fortunate that I caught this when I did lol.

CHKP