Expanded Read Chars Func, Added Pad Length Func
The readNCharsFromStream function now has overloads to support byte vectors, and will only resize the targeted container if it's too small to hold the number of bytes requested.

Additionally, implemented the padLengthTo function, which will round the passed in "length" value up to the nearest multiple of "padTo" (optionally allowing or disallowing zero-length padding).