First Pass at Gecko Value Tracking, Addr Inference
Initial implementation of a system for keeping track of Gecko's Base Address, Pointer Offset, and Gecko Register values, for the purposes of being able to output GCTRM compatible syntax for codetypes which support it.

At present, codetypes for which GCTRM offers custom syntax require their target addresses be explicitly specified; they can't be given relative to the BA or PO values that Gecko's code handler keeps track of on its own. As a result, outputting GCTRM's syntax requires that we be able to calculate the addresses we intend to use, which in turn requires that we be able to track the values of BA and PO. Importantly though, we can't actually *always* calculate those, cuz technically they can be stored to and loaded from RAM; and we obvs can't know what's in RAM at any given moment lol. So, best thing we can do is track changes to those values as closely as possible through Gecko codes that change them, and invalidate the values anytime we change one of their values in a way that'd leave it incalculable to us.

This also necessitates that, for all codetypes which have valid GCTRM syntax, we *need* to implement a hex-and-comment style output solution to use in situations where either BA or PO are invalid at output time. I've implemented such a method for C2 codes with this commit; and will continue to do so for all other codes with GCTRM syntax.