Implemented C0 Code Type
This is the same more or less as the C2 type, it just can't use GCTRM output and shouldn't output an address (these don't use one).

In doing this, I've spruced up non-GCTRM PPC output a little bit; instructions are now aligned to make reading them a bit less of a pain lol.