Implemented Data Embed Detector
A fun little thing I've run into recently is that, in some situations, programmers have made use of some nifty Gecko tricks to store raw data inline with their code. In GCTRM at least, this construction typically goes like this:

	.BA<-DATA_LABEL			# * 4600XXXX 00000000, where XXXX is distance between next line and DATA_LABEL
	.BA->$SOME_ADDR			# * 44000000 XXXXXXXX, where XXXXXXXX is the location to store the pointer to DATA_LABEL
	.GOTO->POST_DATA_LABEL	# * 662000XX 00000000, where XX is the number of lines the embedded DATA takes up
	DATA_LABEL:
	{DATA HERE}				# (XX * 8) bytes of raw hex data, which we need to avoid parsing
	POST_DATA_LABEL:
							# Note: can of course do this using PO instead of BA, the phrase is more or less the same.

We *really* need to avoid trying to parse that hex data as Gecko code, as not only will the output likely not accurately represent the data anyway, but there's a non-zero chance to interpret a line of data as an 06 (String Write) code, which would have us allocate however many bytes are dictated by the following word, which could easily consume an ungodly amount of memory and effectively crash the program (we'd never be able to read enough bytes to escape the byte read loop).

So, as an avoidance measure, I've implemented a simple system for detecting constructions like these so we can avoid the hex data. Basically, every time we hit a 46/4E with an appropriate offset, we store the location in the stream pointed to by that offset. Then, as we process lines, we'll check if we've passed each of the stored locations, and remove
any that have expired. Additionally, each time we parse a GOTO, we'll check if after parsing it we're at one of the locations we suspected was an embed. If it is, we can dump the embedded data, and afterwards discard any suspected locations we passed in dumping the block.

Additionally, fixed an issue where BAPO store codes didn't output their addresses correctly, added support for outputting C0 codetypes in GCTRM's PULSE format (legit just forgot these existed lol), and rolled the logic for dumping un-annotated hex into a function so it could be used both for unrecognized code segments and for data embed dumps.