Added BAPO Store Hint to Data Embed Detector
In some of the test files I'm working with, I've found some situations where the GOTO statement for a given data embed is actually separated from the data itself by some other code. For instance, this example in current P+Ex's SFXExpand.asm:
```
	.BA<-TablePointer
	.BA->$801C75A4
	.GOTO->SkipTable

CodeFlag:
		word[4] 0, 0, 0, 0	# Turned into C0DE, two indexes and a blank space if initialized correctly
TablePointer:
	word[18] |				# Table 1 (0x48 bytes)
	# Table data definitions continue on from here...
```
Here, this CodeFlag region is inserted between the GOTO and the region pointed to by the initial BA store. The current solution failed to catch such situations, since we'd been assuming that the GOTO would always immediately precede the data it's skipping, but there's no real guarantee that this'll be the case (as evidenced here).

So, I've added a second detection criteria: if we're suspecting that an embed is coming because of a 46/4E code, then we BAPO store to an address whose value doesn't involve the BAPO value storing the embed location, and *then* we hit a GOTO *before* the point forecast by the 46/4E: we'll check if that GOTO would take us past the suspected embed location. If so, we'll treat it as a valid embed, and output accordingly.

Additionally, slightly changed the criteria for all embed detections to require that the GOTO that triggers the output have a positive line offset.