Ensured That Unrecognized PPC is Still Output
Previously, things were set up such that unrecognized PPC hex weren't output as part of converted GCTRM style HOOK/PULSE codes. This didn't occur to me as an issue previously, since every valid instruction in the Broadway standard should be supported already; but I've learned recently (through the current iteration of the common2.pac code) that sometimes people embed raw hex as part of their assembly code, same as they do for Gecko codes, and we need to make sure we output those.

This change ensures that happens, but with the caveat that I don't have as surefire a way to identify these as I did for the Gecko version, so embedded data sometimes gets incorrectly output as PPC code. It seems that these embeds are done by just branching past the data itself, but people unconditionally branch past actual code all the time, so I can't easily use that as my only criteria. Not sure how I'm going to fix this, or if I easily can in the first place; it's a dilemma for sure.

And, lastly, changed the layout definitions for the PPC compare instructions to make double word comparisons illegal entirely. This is technically out of alignment with the standard, but since the comparison instructions as I've implemented them are strictly cmpw (not cmpd), accepting hex with the double word bit set could lead to situations where the output conversion doesn't actually match the original hex. This can particularly cause problems when embedded data gets translated into a compare instruction (as in the common2 code, lol); as this changes the hex that the conversion compiles to; which in that case breaks the string. Plus, this project doesn't support any 64-bit instructions anyway; so there's that too.