Slight Adjustments to String Write Output
The ellipses printed when a string spans multiple lines are now separated from the actual string content by a space. Additionally, if a string is started on a new line (eg. for the first string in a String Write), it'll be spaced such that, if it spans multiple lines, the starts of each string segment will line up vertically.