Fixed Broken Parsing on 0xFX Codetypes
The 0xEX codetypes are special actually in that they aren't allowed to use the BAPO bit, which keeps all their codes in the 0xEX range, separate from 0xFX codes. I wasn't accounting for this in the codetype parser, so for example a code like 0xF3 would actually be interpreted erroneously as a 0xE2 code with the BAPO bit set. This has been fixed; 0xF codes no longer lose their BAPO bit when parsing their codetypes.