Implemented F2 Codetype
This required a bit of a reworking of how exactly the output function for C0 and C2 codes worked, since the underlying ASM printing logic is the same between them all. Also, the conversion function used by those three codetypes has been renamed to "geckoASMOutputodeConv" to reflect that it applies to all of them, not just C2.