Implemented F6 Codetype
Also made some slight adjustments to the dumpUnannotatedHexToStream function so that it now uses printStringWithComment to print its first line and comment; which should ensure that its output is always properly in-line with all other comments. Additionally, ensured that if no comment is specified to the printStringWithComment function, it won't try to print one.

Also, this should be the very last Gecko Codetype that I needed to implement; I think we're finally done!