Ensured F2 Codetype Properly Uses PO
Previously, I'd forgotten to account for the fact that the F2 codetype actually doesn't check whether to use PO instead of BA through the usual BAPO bit (since the top nibble is already 0xF). Instead, it uses a second codetype (F4) to enable this functionality. This is now properly accounted for.

Also, fixed an issue where I was zeroing out the higher bits of the length number too early, which prevented the rest of the relevant settings for the codetype from being noted properly.