Implemented Loop and Goto Tracking
Loop tracking for the moment just allows us to indent and un-indent our code comments, which helps to break things up a bit in converted files and just overall helps clarity. Also makes it a lot clearer when a repeat never closes, as everything after it ends up improperly indented. In the future, I may use this to also enable proper GCTRM output for RAM write codes, though I'm still a bit hesitant to do that until I also definitively figure out If codes and Code Execution status (if BAPO gets set in a conditional block, I can no longer know its value, so need to track that).

Other thing is Goto tracking, the primary purpose of which currently is to keep track of whether or not we're currently in a block of code that can be skipped by a Goto. This is currently used to fuel what I think is the last data embed detection case; where we just unconditional Goto past some data straight-up, no moving values into BA or storing anything. Essentially, I've stipulated that the only case under which we may validly assume that a Goto block like this is marking a data block is if we never have a situation where some other Goto would branch *past* a potential Embed Goto and into the data region the embed is supposed to be skipping; as they'd imply that the skipped content is meant to be executable code.