Significant Optimizations to HUD Color Codes
A significant portion of the actual "meat" in the old color changer codes was dedicated purely to filtering out CLR0s that weren't meant to be affected by the codes, which previously we were doing based on names. While that *worked*, filtering that way (even in the abridged, simplified manner I was doing it) just kinda naturally results in rather large code, which is problematic both for the reason that GCTs are space-limited by nature, *and* that expanding this functionality to include any other files (for whatever reason that might be desired) would've caused the code to grow; as we'd have needed to add new cases for any new files.

With these changes though, we're moving away from the old name-checking model to a new activator checking one. CLR0s have a field called "Original Path" in BrawlBox, which as far as I can tell is entirely unused by the game itself. So, I've commandeered it for my own purposes: marking CLR0s that need to be affected by these codes! Setting a CLR0's "OriginalPath" value to "lCLR" now triggers the code, allowing CLR0s to be redirected via the code menu!