Optimized Away Hand Color Codes
In testing, I came across an issue where Eon's "Disable CSS Hand resizing and Hand size based on roster size (BEX) v1.1" code was overwriting a float constant used by the code that was determining hand colors. In the case I was experiencing, rather than that constant being 1.0, it was 0.9, which caused things to round down when converting to integers for my own purposes. To fix this, I've overwritten the call for the modified float with a call to an unmodified one slightly later in memory, which fixes everything.

Additionally, while investigating the above, it occurred to me that, technically, both hand codes *and* the "HUD Color Changer" code both ultimately end up calling the CLR0 "setFrame" function. That, along with the more efficient filtering made possible by the "OriginalAddress" field made it practical to just roll these three codes into a single hook in "setFrame", which has now been done. Saves a pretty decent amount of code space too: after this round of changes the whole compiled thing is about 20% smaller than it was even after the "OriginalAddress" optimizations.