Fixed Inaccurate Branch Prediction Mark Output
The version of the PPC reference manual I was using turns out to actual have been for a later version of the standard than that used for the Wii! As a result, the method I was using for determining appropriate branch prediction marks was incorrect for the target platform, and has been fixed. Big thanks to Gaberboo for pointing out the issue!