Moved ASM Output Prep, Minor Logging Additions
It struck me as out of place to have the notice about parsing the symbol map file before everything else, which made me realize that really there's no reason these functions shouldn't be down near the actual calls for generating the ASM file; and, so, they've been moved. Additionally, added a note about whether or not the ASM file was successfully generated to the log output.