Minor Adjustment to BCLR, BCCTR Output Logic
This ultimately shouldn't affect anything for now, as currently I've elected not to explicitly append the negative branch predictor ("-") when the y-bit isn't set, but if in the future I were to choose to, the previous version would have had the link flag suffix ("l") appended *after* the branch predictor, not before it as intended. This is just a pre-emptive addressing of that issue; it shouldn't meaningfully affect output at all, currently.