Fixed BC Branch Prediction Mark/AA,LK Ordering
Same fix I did for BCLR and BCCTR output functions, somehow missed it in this one lol. Without this, in the event that a branch prediction mark was necessary on a BC instruction which also needed the LK or AA suffixes, those suffixes would've come after the branch prediction mark.

Also had the order of those suffixes backwards before, that shouldn't happen anymore either lol.