Reworked Shield/Death Plume Code
Moves to a new approach where, rather than trying to intercept the actual call to the CLR0 itself, we instead intercept the call to IC-Basic[21029], which is how Shields and Death Plumes natively get their colors. Overall simpler, removes the need for the filtering-style approach I was using before, and saves a bit of space code-wise.