Archiving Test for New Selchar Recolor Approach
So, this was an idea that I had that didn't ultimately pan out due to a conflict with Sammi's Async RSP Loading stuff. The idea here was that we'd simply replace r3+0x1B0 once with the desired color value, let the function run with the patched value, then set it back to its original value afterwards. While this works, as far as setting the colors correctly, the Async RSP Code relies on the patched value being 4 or lower in order to index into a list of pointers; so any of our added colors would crash the game. Could temporarily un-set that value as we enter setCharPic, but I wanna wait till the plugin is more finalized before I try working around it lol. Will come back to this at some point in the future though, could shrink the code down even further.