Minor Optimizations to Team Mode Caching
Rather than just storing a boolean 1 or 0 to represent whether team mode is on, we now store the offset to the relevant code menu line value, VALUE or DEFAULT. This means we don't have to manually check in each place the value is used which value to load, we can just do it once when the value is updated and use that repeatedly.

Additionally, we now only update the team store value when we enter Classic Mode matches, since that's the only circumstance where we really need it.