Began Re-implementing Results Screen Changers
If we do go through with the new selchar approach, we'd want to get rid of the backplateColorChange and menSelChrElemntChange codes, which is what the Results Screen currently relies upon for its color changers. This is the beginning of some work to re-implement those changers independently from those codes.

Additionally, fixes an issue in the Info.pac CLR0 code would fail when trying to load frames above 4, and properly disabled faded CPU team colors on the results screen.