Removed Results Screen Setup Code
This was a fun idea lol, but at the end of the day, it only actually saves one or two instruction over getting a given line's color the normal way, and we don't use it enough times to justify the upfront cost of setting up the optimization, unfortunately. We'd need to use it at least 5 times to break even, simply not worth it. Premature optimization, interesting idea though.

Also, moved the Results Screen Mark fix into the "getMarkColAnimFrame" function. I missed this before, but the game actually does use param_2 like the player slot as far as checking whether anything is *in* that slot, so passing in values for potentially higher color numbers broke things, and even small numbers broke when only one player was present.