Fixed Color 0 on the Results Screen
The int-to-double conversions used in the places I'm hooking are unsigned (which makes sense, since they're designed to act on a loop iterator, which should be unsigned), so they won't properly convert the -1 Color 0 becomes after I do the subtraction in the HOOK. Just preventing the game from adding 1 afterwards in both cases fixes the issue, and also means I don't have to do the subtraction in the hooks themselves, so ultimately the code stays the same size!

I think, with this, the results screen changer is pretty much finalized! Again, whether or not I actually use it depends on how the selchar code ends up, but overall pretty happy with this.