Optimized MenSelChr Element Code
Main optimization here is getting rid of the case where we skip doing a second lookup if we'd be looking for the same CLR0 a second time. Realistically speaking, it's infrequent that we'd even be in that case, since the way that I've set this system up means that not even the default color values line up with their original indices (eg. red originally was index 0, now it's always index 1). The case just didn't really save much time, and was increasing the size of the code a decent bit, so no reason to keep it around.

Also, since we aren't accounting for this case anymore, we don't need to worry about backing up that r3 value and comparing against it later, so we don't need the instructions backup and restore that either!

Last major optimization, and one I'm more conceptually excited about lmao, is the *super* RLWIMI optimization in here: on top of the usual RLWIMI multiplication-insert optimization, I *also* managed to roll a subtraction into it as well lmfao. We're extremely fortunate that the ASCII value for '0' is 0x30, cuz we can just cut off those 5th and 6th bits to subtract the 0x30, and we can do *that* at the same time we do the RLWIMI move lol. I'm thrilled lmfao, very specific, but very fun optimization.