Added a Toggle for Optional MULLI Optimization
I'm honestly at a bit of an impasse regarding whether or not to actually turn this on or not lol. On the hand, it *does* make things very (very) slightly faster for free, but on the other it maybe violates the expectation that the builder will always spit out the instruction you ask it for. I'm leaving the code for it here, in case I do decide to come back to it, but for now it'll be turned off.