EX_Config XML Can Now Specify CharVersionList
This replacing the option originally present in the Options XML, you can now use the "baseListVersion" attribute of an EX Character Declaration block to choose the base list to which your character declarations will be added.

To facilitate this, implemented a dedicated function for changing the active list, as well as an array which provides name strings for each list preset.