Removed Editable CharList and MAIN_PATH From Options XML
These have been moved into the Config XML, and so will no longer be editable here.

Additionally, swapped the old "name" attribute on the codeMenu node for a proper path attribute, displaying the full absolute path to the menu on the SD Card. To facilitate this, a new function has been implemented for getting the Menu's Absolute path, which is used both for the path attribute in the Options XML and in "Control Codes.cpp" to get the path in the "LoadCodeMenu()" function.