Flattened and Broke Up Config XML Parse Func
The previous iteration of the function was getting pretty deeply indented lol, was kinda annoying to work with, so I've flattened it out by just having us return a bit early if we fail to meet the conditions necessary to actually parse the document's contents.

Additionally, I've split the code for actually adding entries from each declaration block to the menu into their own functions, just to break things up a bit further there as well.