Moved to Dual Config XML System
In order to allow the offline and online versions of this program to utilize different configurations, the program now recognizes the existence of a second config file "Net-EX_Config.xml". When running the Netplay version of the builder, it will first check for that file and use it if it exists, falling back to the normal "EX_Config.xml" file otherwise.