Rewrote Menu Loaded Check to Not Rely on Title
Checking the title made it so that *changing* the title broke the menu, doing away with that.
Instead, we'll check if the pointer to the start of the menu body is present at the Header Loc.
Accomplishes the same thing, allows the title to be changed, and should will never need to be changed by hand!