Slight Adjustments to Config Parsing Log Messages
General adjustments so that, if a section of the config document isn't populated with any input from the user, no log messages are produced about it. This may see additional adjustment in the future as the config system is further refined and tweaked.

Additionally, adjusted menu comments block parser to check whether the flag for deleting the default header comments is set, regardless of whether or not any additional comments are actually set.