Fixed an Oversight in the Menu Load Check Rewrite
Fixed an issue in the rewrite of the menu load check which was preventing the menu from properly closing if you entered a sub-page. This occurred due to using the first word in the Header, which stores a pointer to the active menu page, rather than the second word, which always stores a pointer to the main page (what I was looking for). This has been fixed, the menu now opens and closes as intended.