Non-EX Builds Use "Config.xml", not "EX_Config.xml"
I intend to offer a second set of executables going forwards which use the regular, non-EX P+ config, in case anyone wants to use the stuff I've implemented in this repo for regular P+ builds; and having the config file have "EX" in the names for the input files kinda gave the impression that the builder only worked for EX builds. This does away with that, and should hopefully make it clearer that the builder works for *both* configs!